+++ core/designsystem/src/main/java/com/example/designsystem/component/CowGroupDialog.kt
@@ -1,4 +1,4 @@
-package com.example.designsystem
+package com.example.designsystem.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
