+++ core/designsystem/src/main/java/com/example/designsystem/component/DatePickerTextField.kt
@@ -1,4 +1,4 @@
-package com.example.designsystem
+package com.example.designsystem.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
