+++ core/designsystem/src/main/java/com/example/designsystem/component/MarkButton.kt
@@ -1,4 +1,4 @@
-package com.example.designsystem
+package com.example.designsystem.component
 
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
