+++ feature/home/src/main/java/com/example/home/presentation/CreateMeetingScreen.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.example.common.DateUtil.todayStartOfDayMillis
-import com.example.designsystem.DatePickerTextField
+import com.example.designsystem.component.DatePickerTextField
 import com.example.home.component.Category
 import com.example.home.component.CategoryDropdown
 import com.example.home.component.CowGroupSlider
