+++ feature/home/src/main/java/com/example/home/presentation/EventDetailScreen.kt
@@ -42,8 +42,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.designsystem.CowGroupDialog
-import com.example.designsystem.MarkButton
+import com.example.designsystem.component.CowGroupDialog
+import com.example.designsystem.component.MarkButton
 import com.example.home.R
 import com.example.home.viewmodel.EventDetailUIState
 import com.example.home.viewmodel.EventDetailViewModel
