+++ feature/home/src/main/java/com/example/home/presentation/HomeScreen.kt
@@ -26,7 +26,7 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import androidx.paging.compose.itemKey
 import com.example.data.model.SearchHistory
-import com.example.designsystem.PagingMeetingItem
+import com.example.designsystem.component.PagingMeetingItem
 import com.example.home.component.HomeScreenSearchBar
 import com.example.home.viewmodel.HomeUIState
 import com.example.home.viewmodel.HomeViewModel
