+++ feature/mypage/src/main/java/com/example/mypage/component/MeetingItem.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.designsystem.MarkButton
+import com.example.designsystem.component.MarkButton
 import com.example.model.MyPageEvent
 import com.example.mypage.R
